requirements.txt
# Core Framework
streamlit==1.31.0
python-dotenv==1.0.1

# AI Providers
google-generativeai==0.3.2
openai==1.12.0

# Database
sqlalchemy==2.0.27

# Data Processing & Validation
pydantic==2.6.1
pillow==10.2.0
python-multipart==0.0.9
pytesseract==0.3.13

# Security
bcrypt==4.1.2
pyjwt==2.8.0

# Utilities
python-dateutil==2.8.2
pytz==2024.1

# Visualization
plotly==5.19.0
pandas==2.2.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
